+++ jbm-cluster/jbm-cluster-platform/jbm-cluster-platform-center/src/main/java/com/jbm/cluster/center/service/impl/BaseMenuServiceImpl.java
@@ -103,7 +103,7 @@ public class BaseMenuServiceImpl extends MasterDataServiceImpl<BaseMenu> impleme
         QueryWrapper<BaseMenu> queryWrapper = new QueryWrapper();
         queryWrapper.lambda()
                 .eq(BaseMenu::getMenuCode, menuCode);
-        int count = baseMenuMapper.selectCount(queryWrapper);
+        Long count = baseMenuMapper.selectCount(queryWrapper);
         return count > 0 ? true : false;
     }
